check-s3.sh

+++ check-s3.sh
@@ -9,7 +9,7 @@ do
 	echo "Working on $bucket ..."
     SIZE="$(aws s3 ls s3://$bucket --recursive --summarize | grep Size | awk '{print $3}')"
 
-    if [ $SIZE -gt 20971520 ]
+    if [ $SIZE != "" && $SIZE -gt 20971520 ]
     then
         echo "Emptying $bucket ..."
         aws s3 rm s3://$bucket --recursive
